[kunlunxin] cleanup: remove unused imports (F401) and trailing whitespace
- fused: drop unused _general_module shim imports (hc_split_sinkhorn,
  mhc_pre); module-loading side effect preserved via from-import
- ops: drop unused torch / tl_extra_shim / _FALLBACK_KEYSET imports
  (attention, linalg_lu, diagonal_scatter, _prelu_kernel_backward,
  index_fill)
- ops/__init__: noqa F401 on SpecOpRegistrar re-exports (addbmm,
  linalg_lu, narrow, resize_output)
- trim trailing whitespace and add missing EOF newlines across vendor
  ops/fused files

+++ src/flag_gems/runtime/backend/_kunlunxin/ops/cumprod.py
@@ -378,9 +378,7 @@ def cumprod_strided_row_scan_kernel(
     offs = tl.arange(0, TILE_SIZE)
     mask = offs < N
     acc_dtype: tl.constexpr = get_prod_accum_type(out_ptr.type.element_ty)
-    x = tl.load(inp_ptr + row_off + offs * stride_n, mask=mask, other=1).to(
-        acc_dtype
-    )
+    x = tl.load(inp_ptr + row_off + offs * stride_n, mask=mask, other=1).to(acc_dtype)
     r = tl.cumprod(x, 0)
     tl.store(
         out_ptr + row_off + offs * stride_n,
@@ -524,6 +522,6 @@ def cumprod_(inp, dim, *, dtype=None):
         # Inclusive prefix product over an axis of length 1 is the identity,
         # so the in-place op needs no work at all.
         return inp
-        
+
     cumprod_wrapper(inp, dim, inp.dtype, out=inp)
     return inp
